+++ src/backend/controllers/base_profiles.controllers.js
@@ -1,16 +1,15 @@
 const asyncHandler = require("express-async-handler");
-const { Announcement } = require("../models/announcement");
 const {Base_profile} = require("../models/base_profile");
-const {User} = require("../models/user");
-const {Club} = require("../models/club");
-
+const { deleteAnnouncementByUser_Id } = require("./announcements.controllers");
+const { deleteUserByBase_profileId } = require("./users.controllers");
+const { deleteClubByBase_profileId } = require("./clubs.controllers");
 
 const deletebase_profileById = asyncHandler(async(req,res)=>{
   const base_profileid = req.params.id ;
   try{
-    await Club.destroy({where: {base_profile_id: base_profileid}});
-    await User.destroy({where: {base_profile_id: base_profileid}});
-    await Announcement.destroy({where: {user_id: base_profileid}});
+    deleteAnnouncementByUser_Id(req,res) ;
+    deleteUserByBase_profileId(req,res) ;
+    deleteClubByBase_profileId(req,res) ;
     await Base_profile.destroy({where: {id: base_profileid}});
 
     res.status(200).json({ success: true, message: 'profile deleted successfully'});
