+++ src/backend/controllers/users.controllers.js
@@ -1,7 +1,7 @@
 const asyncHandler = require("express-async-handler");
 const { User } = require("../models/user");
 const { deleteCouncilByUser_Id } = require("./council.controllers");
-const { deleteMembershipByUser_Id } = require("./membership.controllers");
+const { deleteMembershipByUser_Id } = require("./memberships.controllers");
 const { deleteThreadByUser_Id } = require("./threads.controllers");
 const { deleteCommentByUser_Id } = require("./comments.controllers");
